fix: as receitas estao sendo cadastradas de forma correta

+++ src/main/java/org/example/nutricomebh/Quantidade/QuantidadeDTO.java
@@ -6,6 +6,8 @@ import lombok.Data;
 import lombok.NoArgsConstructor;
 import org.example.nutricomebh.ItemReceita.ItemModel;
 
+import java.math.BigDecimal;
+
 @Data
 @AllArgsConstructor
 @NoArgsConstructor
@@ -13,6 +15,6 @@ public class QuantidadeDTO {
 
 
     private Long id;
-    private double quantidade;
+    private BigDecimal quantidade;
     private ItemModel item;
 }
